Develop

+++ src/utils.py
@@ -1,16 +1,12 @@
 import os
 import glob
 import yfinance as yf
-from datetime import datetime, timedelta
 
-def import_data_yf(symbol):
-    end_date = datetime.today().strftime('%Y-%m-%d')
-    start_date = (datetime.today() - timedelta(days=365)).strftime('%Y-%m-%d')
-    
+def import_data_yf(symbol, start_date, end_date):
+    """Download financial data using yfinance."""
     try:
         df = yf.download(symbol, start=start_date, end=end_date, interval='1d')
         df.columns = ["open", "high", "low", "close", "adj close", "volume"]
-        df = df.drop(columns=["adj close", "volume"])
         df.index.name = "time"  
     except Exception as e:
         print(f"An error occurred: {e}")
@@ -19,15 +15,19 @@ def import_data_yf(symbol):
     return df
 
 def clear_directory(path):
-    """Deletes all files in the specified directory."""
-    files = glob.glob(os.path.join(path, '*'))
-    for f in files:
-        try:
-            os.remove(f)
-            print(f"Deleted: {f}")
-        except Exception as e:
-            print(f"Error deleting file {f}: {e}")
+    """Clear all contents of the output directory."""
+    if os.path.exists(path):
+        files = glob.glob(os.path.join(path, '*'))
+        for f in files:
+            try:
+                os.remove(f)
+                print(f"Deleted: {f}")
+            except Exception as e:
+                print(f"Error deleting file {f}: {e}")
+    else:
+        print(f"Directory {path} does not exist. No files to delete.")
 
-def create_directory(path):
-    """Ensures the directory exists, creating it if necessary."""
-    os.makedirs(path, exist_ok=True)
+def create_directory(output_dir):
+    """Create the output directory if it does not exist."""
+    os.makedirs(output_dir, exist_ok=True)
+    print(f"Directory {output_dir} created or already exists.")
